Add explicit record script, remove irb hook
We'll be starting with pry since we got it installed yesterday. We could
implement a similar .pryrc hook, but pry restores readline history, so each dump
would have all the commands from previous sessions, which isn't terribly useful.

Using this approach, we have the full history, including the produced output.